constants.py moved to utils
