+++ tests/test_new_app_scenario.py
@@ -6,10 +6,10 @@ from contextlib import redirect_stdout
 
 import pytest
 
-from constants import BUILD_AGENT_SUCCESSFUL, TEST_AGENT_SUCCESSFUL
 from scenarios.new_app.new_app_scenario import NewAppScenario
 from tests.test_utils import setup_test
 from utils.config import Config
+from utils.constants import BUILD_AGENT_SUCCESSFUL, TEST_AGENT_SUCCESSFUL
 from utils.misc import to_os_path
 from utils.tee import Tee
